refactor: rename mcpVerifier.ts to mcpClient.ts for better clarity
Renamed files to eliminate naming confusion between the core MCP client
logic and the command handler:

- src/core/mcpVerifier.ts → src/core/mcpClient.ts
- tests/core/mcpVerifier.test.ts → tests/core/mcpClient.test.ts

Updated imports in:
- src/commands/apply.ts
- src/commands/verifyMcp.ts
- tests/commands/verifyMcp.test.ts

The new naming makes the architecture clearer:
- mcpClient.ts: Low-level MCP client operations and verification logic
- verifyMcp.ts: High-level CLI command orchestration
