+++ src/core/mcpClient.ts
